requirements.txt
requests
openai
pydub
python-dotenv
